Delete or move config files
