+++ packages/yoastseo/index.js
@@ -5,7 +5,7 @@ import * as helpers from "./src/helpers/_todo";
 import * as markers from "./src/markers";
 import * as interpreters from "./src/scoring/interpreters";
 import * as researchers from "./src/languageProcessing";
-import * as config from "./src/config/_todo";
+import * as config from "./src/config";
 
 import App from "./src/app";
 import Assessor from "./src/scoring/assessor";
