+++ packages/yoastseo/src/app.js
@@ -1,4 +1,4 @@
-import "./config/_todo/config.js";
+import "./config/config.js";
 import SnippetPreview from "./snippetPreview/snippetPreview.js";
 
 import { defaultsDeep } from "lodash-es";
