--- packages/yoastseo/src/config/_todo/removalWords.js
@@ -1,10 +1,0 @@
-/** @module config/removalWords */
-
-/**
- * Returns an array with words that need to be removed
- *
- * @returns {array} removalWords Returns an array with words.
- */
-export default function() {
-	return [ " a", " in", " an", " on", " for", " the", " and" ];
-};
