+++ packages/yoastseo/src/config/config.js
