+++ packages/yoastseo/src/config/index.js
