+++ packages/yoastseo/src/scoring/assessments/seo/MetaDescriptionLengthAssessment.js
@@ -1,6 +1,6 @@
 import { merge } from "lodash-es";
 import Assessment from "../assessment";
-import Config from "../../../config/_todo/config";
+import Config from "../../../config/config";
 import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
 import AssessmentResult from "../../../values/AssessmentResult";
